+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 - ComponentBuilder implementations, lambdas replaced with static classes to prevent memory leaks.
 ### is.codion.framework.i18n
 - FrameworkMessages, confirm update and confirm insert messages improved.
+- FrameworkMessages, no search results message improved.
 ### is.codion.swing.framework.ui
 - EntitySearchField.selectorFactory() now Value based.
 - EntitySearchField.searchIndicator() now Value based.